Crate problem109

Problem 109 (Darts)

Solution detail

This is an easy application of brute force - simply iterate over all combinations of two darts (including a miss) plus a double, and count how many have a total less than 100.

Be careful when iterating over the choice for the first two darts that the same two darts in different orders are not counted twice.

Functions

answer
Solve the problem, returning the answer as a String

main

solve
Find the number of darts checkouts for scores less than the given limit.